requirements.txt
fastapi
uvicorn
pandas
joblib
scikit-learn
pyjwt
streamlit
requests
pydantic